adding cron job configs
