+++ .gitignore
@@ -1,2 +1,5 @@
 node_modules
 package-lock.json
+.dev.env
+.prod.env
+.env
