+++ stakepool-cron/Mainnet/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "stakepool-oro",
+  "version": "1.0.0",
+  "description": "",
+  "main": "handler.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@taquito/signer": "^8.0.6-beta.0",
+    "@taquito/taquito": "^8.0.6-beta.0",
+    "axios": "^0.21.1",
+    "express": "^4.17.1"
+  }
+}
