+++ stakepool-cron/Testnet/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "stakepool-cron-testnet",
+  "version": "1.0.0",
+  "description": "stakepool-cron-testnet",
+  "main": "handler.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node handler.js"
+  },
+  "author": "aditya@tezsure.com",
+  "license": "ISC",
+  "dependencies": {
+    "@taquito/signer": "^8.0.6-beta.0",
+    "@taquito/taquito": "^8.0.6-beta.0",
+    "axios": "^0.21.1",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1"
+  },
+  "devDependencies": {}
+}
