+++ stakepool-cron/Testnet/serverless.yml
@@ -1,0 +1,16 @@
+service: stakepool-cron
+frameworkVersion: '2'
+useDotenv: true
+provider:
+    name: aws
+    region: ap-south-1
+    runtime: nodejs12.x
+    lambdaHashingVersion: 20201221
+
+functions:
+    pingContract:
+        handler: handler.pingContract
+        events:
+            - schedule:
+                  rate: cron(*/2 * * * ? *)
+                  enabled: true
